Added `de/deserializer` module
